requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.1.1
scikit-learn==1.2.2
pandas==3.0.2
numpy==1.26.4
joblib==1.2.0
PyMySQL==1.0.2
lightgbm==4.5.0
xgboost==2.1.1
matplotlib==3.9.2
Pillow==10.4.0
python-docx==1.1.2
python-pptx==1.0.2
torch==2.5.1
sentence-transformers==3.3.1
